+++ resources/js/app.js
@@ -1,14 +1,17 @@
-import { Application } from 'stimulus';
+import { Application, Controller } from 'stimulus';
 import { definitionsFromContext } from 'stimulus/webpack-helpers';
+import { $, jQuery } from 'jquery';
 
 // remove
 $.fn.select2.defaults.set('theme', 'bootstrap');
 
-const jq = require('jquery');
 
-global.$ = jq;
-global.jQuery = jq;
+window.$ = $;
+window.jQuery = jQuery;
+
 
 window.application = Application.start();
+window.Controller = Controller;
+
 const context = require.context('./controllers', true, /\.js$/);
 application.load(definitionsFromContext(context));