+++ resources/js/controllers/screen/tabs_controller.js
@@ -1,7 +1,6 @@
-import {Controller} from 'stimulus';
+import { Controller } from 'stimulus';
 
 export default class extends Controller {
-
     /**
      *
      */