+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.1.4] - 2018-12-09
+
+### Added
+- Global variable `Controller`
+
+### Removed
+- Unused Image package dependency
+- Unused `JsonRelationsTrait`
+
 ## [3.1.3] - 2018-12-06
 
 ### Deprecated